+++ atlas.com/npc/data/etc/model.go
@@ -3,7 +3,7 @@ package etc
 type Model struct {
 	id        uint32
 	price     uint32
-	unitPrice uint32
+	unitPrice float64
 	slotMax   uint32
 }
 
@@ -15,7 +15,7 @@ func (m Model) Price() uint32 {
 	return m.price
 }
 
-func (m Model) UnitPrice() uint32 {
+func (m Model) UnitPrice() float64 {
 	return m.unitPrice
 }
